% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asymmTree.R
\name{asymmTree}
\alias{asymmTree}
\title{asymmTree}
\usage{
asymmTree(n)
}
\arguments{
\item{n}{The number of tips for the tree to have.}
}
\description{
Simulates a completely unbalanced tree with n tips, and branch lengths calculated according to Grafen.
}

