% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchChangeSim.R
\name{branchChangeSim}
\alias{branchChangeSim}
\title{branchChangeSim
Simulate the expected change along a branch of length branchtime.}
\usage{
branchChangeSim(branchtime = 100, direction = "non", startval = 0,
  var = 0.2, increment = 1)
}
\description{
branchChangeSim
Simulate the expected change along a branch of length branchtime.
}
