% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladeChangeSim.R
\name{cladeChangeSim}
\alias{cladeChangeSim}
\title{cladeChangeSim
Simulates the expected change along all branches of a defined clade.}
\usage{
cladeChangeSim(tree, node, directional = "non", startval = 0, sigsq,
  plot = FALSE)
}
\description{
cladeChangeSim
Simulates the expected change along all branches of a defined clade.
}

