% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{createCountsTable}
\alias{createCountsTable}
\title{#' A bunch of internal functions for hfgr, mostly for the localscalar PP.}
\usage{
createCountsTable(extree, meanbl)
}
\description{
#' A bunch of internal functions for hfgr, mostly for the localscalar PP.
}

