% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovProbs.R
\name{getCovProbs}
\alias{getCovProbs}
\title{getCovProbs
A function to return the probabilities of independent or dependent at each node 
after fitting the covarion discrete model.
probgraphs will make a pdf of the overlap in probabilities at each node.}
\usage{
getCovProbs(log, tree, ignore50 = FALSE, probgraphs = FALSE)
}
\description{
getCovProbs
A function to return the probabilities of independent or dependent at each node 
after fitting the covarion discrete model.
probgraphs will make a pdf of the overlap in probabilities at each node.
}

