% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescs.R
\name{getDescs}
\alias{getDescs}
\title{getDesc}
\usage{
getDescs(tree, node, nds = NULL)
}
\arguments{
\item{tree}{A tree of class phylo}

\item{node}{Either a single node, or a vector of tip labels}
}
\description{
A function to get all descendant nodes from a given node, or vector of tip labels.
}
