% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLparams.R
\name{getMLparams}
\alias{getMLparams}
\title{getMLparams}
\usage{
getMLparams(logfile)
}
\arguments{
\item{logfile}{The name of the logfile that needs to be examined, or vector of logfile names.}
}
\description{
A function that will return the ML estimates from the logfile of an ML bayestraits run.
}
