% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPL.R
\name{getPL}
\alias{getPL}
\title{getPL
Gets the path length from the root to the position of a particular node.}
\usage{
getPL(tree, startnode = NA, node)
}
\description{
getPL
Gets the path length from the root to the position of a particular node.
}
