% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTipNames.R
\name{getTipNames}
\alias{getTipNames}
\title{getTipNames}
\usage{
getTipNames(tree, node)
}
\arguments{
\item{tree}{A tree of class phylo.}

\item{node}{The node number of interest.}
}
\description{
A function to get the names of the descendant tips from a given node of a tree.
}

