% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotShareLeg.R
\name{ggplotShareLeg}
\alias{ggplotShareLeg}
\title{ggplotShareLeg}
\usage{
ggplotShareLeg(...)
}
\arguments{
\item{...}{Some number of ggplot objects.}
}
\value{
A multiple panel plot of plots with a common legend.
}
\description{
A function to plot multiple ggplot objects with a shared legend.
HFG: from https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
p1 <- qplot(carat, price, data=dsamp, colour=clarity)
p2 <- qplot(cut, price, data=dsamp, colour=clarity)
p3 <- qplot(color, price, data=dsamp, colour=clarity)
p4 <- qplot(depth, price, data=dsamp, colour=clarity)
ggplotShareLeg(p1, p2, p3, p4)
}

