% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gLegend.R
\name{gLegend}
\alias{gLegend}
\title{gLegend}
\usage{
gLegend(a.gplot)
}
\arguments{
\item{a.gplot}{A ggplot object.}
}
\description{
A function to strip the legend from a ggplot object so that it can then be resued. For
example, you might want to plot 6 figures on one page with the same legend. This function
can take the legend from one of the plots, and then plot it later on after the 6 plots
are plotted with no legend. Normally used as part of ggplotSharedLegend.
HFG: From https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
\examples{
leg <- gLegend(my_plot)
}

