% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/hfgr-desc.R
\docType{package}
\name{hfgr}
\alias{hfgr}
\alias{hfgr-package}
\title{A package of personally useful functions.}
\description{
A package of personally useful functions.
}
\author{
Henry Ferguson-Gow
}

