% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/isDefined.R
\name{isDefined}
\alias{isDefined}
\title{is.defined}
\usage{
isDefined(x)
}
\arguments{
\item{x}{The variable or object you want to test the existence of.}
}
\description{
Internal function that tests if a variable is defined (i.e. is not NULL).
}

