% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lmEq.R
\name{lmEq}
\alias{lmEq}
\title{lmEq}
\usage{
lmEq(df)
}
\arguments{
\item{df}{data frame with x and y as the x and y variables.}
}
\value{
a string that can be used as a label for ggplot.
}
\description{
A function to generate the linear model equation and r-squared for labelling plots of
two variables against each other - e.g. branch lengths recoverd from rj bayes traits.
}
\examples{
lmEq(data.frame(x = dat$og, y = dat$rj))
}

