% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRJ.R
\name{loadRJ}
\alias{loadRJ}
\title{loadRJ}
\usage{
loadRJ(logfile, burnin = 0, thinning = 1)
}
\arguments{
\item{logfile}{The name of the logfile of the BayesTraits analysis.}
}
\value{
A list containing the taxa translation table, all possible subtrees a scalar can occur on, and a data frame of the rj model configuration.
}
\description{
Returns the full mcmc object from a BayesTraits log file. This
is used inside plot functions and so on, but might be useful for
other MCMC manipulations and so on.
}
