% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRJdisc.R
\name{loadRJdisc}
\alias{loadRJdisc}
\title{loadRJdisc}
\usage{
loadRJdisc(logfile, burnin = 0, thinning = 1)
}
\arguments{
\item{logfile}{The name of the logfile of the BayesTraits analysis.}

\item{burnin}{burning}

\item{thinning}{Thin the sample by ever n sample.}
}
\value{
A list containing the taxa translation table, all possible subtrees a scalar can occur on, and a data frame of the rj model configuration.
}
\description{
Returns the full mcmc object from a BayesTraits discrete model log file. This
is used inside plot functions and so on, but might be useful for
other MCMC manipulations and so on.
}

