% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localEB.R
\name{localEB}
\alias{localEB}
\title{localEB}
\usage{
localEB(tree, node, a, rescale = TRUE)
}
\arguments{
\item{tree}{A tree object of class phylo}

\item{node}{The node that the transformation should be applied at (includes all branches and nodes downstream of here).}

\item{a}{The EB parameter for the transformation.}

\item{rescale}{Rescale the tree after transformation? Defailts to TRUE.}
}
\description{
A function to transform a tree according to the exponential change/Early Burst model of Harmon et al 2010
}
