% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localKappa.R
\name{localKappa}
\alias{localKappa}
\title{localKappa}
\usage{
localKappa(tree, node, kappa, rescale = TRUE)
}
\arguments{
\item{tree}{A tree of class phylo.}

\item{node}{A node number describing the clade to be transformed.}

\item{kappa}{The multiplier that rate is increased by in the clade specified by node.}
}
\description{
A function that transforms branch lengths according to Kappa.
}
