% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRate.R
\name{localRate}
\alias{localRate}
\title{localRate}
\usage{
localRate(tree, node, scalar)
}
\arguments{
\item{tree}{A tree of class phylo.}

\item{node}{A node number describing the clade to be transformed.}

\item{scalar}{The multiplier that rate is increased by in the clade specified by node.}
}
\description{
A function that transforms branch lengths according to a scalar.
}
