% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localscalarPP2.R
\name{localscalarPP2}
\alias{localscalarPP2}
\title{localscalarPP2}
\usage{
localscalarPP2(rjlog, rjtrees, tree, burnin = 0, thinning = 1,
  meanbranches = TRUE, ratestable = TRUE)
}
\arguments{
\item{rjlog}{The RJ output of the run - typically suffixed with .VarRates.txt}

\item{tree}{The tree the analysis was run on}

\item{burnin}{The burnin (if required) for the mcmc (generally worked out from the other logfile)}

\item{thinning}{Thinning parameter for the MCMC output - again, worked out from the raw MCMC output logfile.}

\item{returnscales}{Return full distributions of all scalar values that hit each node?}

\item{returnorigins}{Return a list of the values of each scalar that originates at each node per iteration?}
}
\description{
A function that takes the output of a kappa, lambda, delta, VRates etc. RJ bayesTraits run and runs post-processing on it.
}
