% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{makeTransparent}
\usage{
makeTransparent(..., alpha = 0.5)
}
\arguments{
\item{alpha}{The alpha required.}
}
\description{
Some function from stack exchange that makes a colour transparent according to a given alpha.
}

