% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/massLh.R
\name{massLh}
\alias{massLh}
\title{massLh}
\usage{
massLh(prefix)
}
\arguments{
\item{prefix}{The prefix of the log files. Needs to be followed, in the logfiles file names, with a number, then ".txt.starter.log" or ".txt.true.log"}
}
\description{
A function that will get the likelihoods for all the starter and true log files
in a directory, and return a data frame of those values for future comparison.
}

