% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanBranches.R
\name{meanBranches}
\alias{meanBranches}
\title{meanbranches}
\usage{
meanBranches(reftree, trees, burnin = 0, thinning = 1)
}
\arguments{
\item{reftree}{A tree that provides the reference topology (ideally the tree the analysis was run on)}

\item{trees}{The posterior sample of stretched trees from which you want the mean branch lengths.}
}
\description{
Calculate the mean branch lengths from a BayesTraits RJ analysis posterior, when topology is fixed.
}

