% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{A list of plot objects (as opposed to individual objects)}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
HFG: From http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_%28ggplot2%29/.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\examples{
multiplot(plot1, plot2, plot3, plot4, cols = 2)
all.plots <- list(plo1, plot2, plot3, plot4)
multiplot(plotlist = all.plots, cols = 2)
}
