% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylo.R
\name{plotPhylo}
\alias{plotPhylo}
\title{plotPhylo}
\usage{
plotPhylo(tree, tips = FALSE, nodes = NULL, scale = TRUE, cex = 1,
  main = "", edge.cols = "black", edge.width = 1)
}
\arguments{
\item{tree}{An object of class phylo}

\item{tips}{Logical - show tip labels or not?}

\item{nodes}{TRUE shows all node labels, FALSE supresses node labels (default), or a vecotr of which nodes to highlight.}

\item{scale}{Logical - show the scale bar, or not?}

\item{cex}{Scaling factor for tip labels and node labels - universal.}

\item{main}{A string that will be the title on the plot (if desired)}

\item{edge.cols}{A vector of edge colours, if you want the edges to be coloured.}
}
\description{
A silly little function that makes plotting a tree without tip labels and with
an axis easier and quicker.
}

