% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{plotTree}
\usage{
plotTree(tree, tips = FALSE, nodes = FALSE, scale = TRUE, cex = 1)
}
\arguments{
\item{tree}{An object of class phylo}

\item{tips}{Logical - show tip labels or not?}

\item{nodes}{Logical - show nodel labels, or nor?}

\item{scale}{Logical - show the scale bar, or not?}

\item{cex}{Scaling factor for tip labels and node labels - universal.}
}
\description{
A silly little function that makes plotting a tree without tip labels and with
an axis easier and quicker.
}

