% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocZ.R
\name{posthocZ}
\alias{posthocZ}
\title{posthocZ}
\usage{
posthocZ(tree, rjtrees, nsim = 100)
}
\arguments{
\item{tree}{The time tree, an object of class phylo, or the filename of the tree to load.}

\item{rjtrees}{The posterior of trees in multiphylo form, or the filename of the posterior.}

\item{nsim}{The number of simulations per tree in the posterior.}
}
\description{
Simulates a trait over the posterior sample of trees, and then calculates z scores.
}
