% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readDiscMaps.R
\name{readDiscMaps}
\alias{readDiscMaps}
\title{readDiscMaps}
\usage{
readDiscMaps(logfile, tree, burnin = 0, thinning = 1, returnpies = FALSE)
}
\description{
A function to read in and interpret the discrete/independent maps that result from model 11.
}

