% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPP.R
\name{readPP}
\alias{readPP}
\title{readPP}
\usage{
readPP(logfile)
}
\arguments{
\item{logfile}{The recorded output of VarPP-v1}
}
\description{
A function that reads the output of the variable rates post-processor (VarPP-v1). This is the output that VarPP-v1 prints to screen, so that needs to be written to a text file first.
}

