% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSimInfo.R
\name{readSimInfo}
\alias{readSimInfo}
\title{readSimInfo
A function that will read the simulation details from Andrews conDatSim function.}
\usage{
readSimInfo(filename, tree)
}
\arguments{
\item{filename}{The filename of the simulation details file.}

\item{tree}{The tree the data were simulated over}
}
\description{
readSimInfo
A function that will read the simulation details from Andrews conDatSim function.
}

