% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJbranchScalar.R
\name{RJbranchScalar}
\alias{RJbranchScalar}
\title{RJbranchScalars
A function that takes the output of a kappa, lambda, delta model and produces a summary of
which branches were scaled, how much, and what value that scalar has, and a tree with the mean branch
lengths of the trees sampled during the analysis.}
\usage{
RJbranchScalar(rjlog, rjtrees, tree)
}
\arguments{
\item{rjlog}{The logfile from the RJ portion of the BT analysis.}

\item{rjtrees}{The filename of the posterior sample of trees from the analysis.}

\item{tree}{The tree on which the analysis was performed.}
}
\description{
RJbranchScalars
A function that takes the output of a kappa, lambda, delta model and produces a summary of
which branches were scaled, how much, and what value that scalar has, and a tree with the mean branch
lengths of the trees sampled during the analysis.
}

