% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantTransformation.R
\name{significantTransformation}
\alias{significantTransformation}
\title{significantTransformation}
\usage{
significantTransformation(PP, scalar, measure = "median", threshold = 0)
}
\arguments{
\item{PP}{The psotprocessor (localscalrPP) output.}

\item{scalar}{The scalar to find and plot from the post processor - delta/lambda/kappa/node/branch}

\item{measure}{Median (default), mean or mode scalar.}

\item{threshold}{Threshold of probability in posterior to display deltas for, defaults to zero (i.e. shows all deltas shaded proportionally to the posterior probability)}
}
\description{
Transforms a tree according to only the scalars present above a certain threshold.
}

