% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDepIndep.R
\name{simDepIndep}
\alias{simDepIndep}
\title{simDepIndep}
\usage{
simDepIndep(itts, treesize, mintax, maxtax, clades, base_mode, base_rate = 2)
}
\arguments{
\item{itts}{The number of trees and datasets to generate.}

\item{treesize}{The size, in terminal taxa, of the tree to be simulated.}

\item{mintax}{The minimum size for the clades that vary from the background model.}

\item{maxtax}{The maximum size for the clades that vary from the background model.}

\item{clades}{The number of clades to simulate the opposite model in.}

\item{base_mode}{The model for the whole tree that the clades vary from. E.g. if "dependent" then the tree will be a dependent model, with n clades evolved as independent. If "independent" than vice versa.}

\item{base_rate}{The model defaults to a "base transition rate" of 2 changes per mean path length, this changes that.}
}
\description{
A function that simulates either dependent or independent evolution between two binary
character over a tree, and then takes some number of clades of a specific size, and simulates
dependent or independent evolution (whichever the rest of the tree isn't) within those clades.
The tree and data are simulated, and the Q matrices for the two traits are fixed for both 
and independent evolution (at the moment).
}
