% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOverTrees.R
\name{simOverTrees}
\alias{simOverTrees}
\title{simOverTrees
This function will simulate the expected change over a vector of named trees. It calls cladeChangeSim a bunch of times.}
\usage{
simOverTrees(tree, reftrees, nsim, node, sigsq, direction, increment = NULL)
}
\arguments{
\item{tree}{The original tree.}

\item{reftrees}{A vector of trees that differ to the original tree on which to simulate the change.}

\item{nsim}{The number of simulations per tree.}

\item{node}{A node from which to conduct the simulation.}

\item{sigsq}{A single sigma squared for the simulation, or a vector of them in the order of the trees.}

\item{direction}{Positive selection ("pos"), negative selection ("neg"), or non directional ("non")?}

\item{increment}{The size of the time interval each branch is broken down into, if NULL is calculated automatically, but may not be comparable between runs.}
}
\description{
simOverTrees
This function will simulate the expected change over a vector of named trees. It calls cladeChangeSim a bunch of times.
}

