% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuDatLambda.R
\name{simuDatLambda}
\alias{simuDatLambda}
\title{simuDatLambda}
\usage{
simuDatLambda(tree, node, sig, w, dat = NULL)
}
\arguments{
\item{tree}{A tree of class 'phylo'}

\item{sig}{The standard deviation for the Brownian Motion process.}

\item{w}{The weight parameter. 1 = strong signal, 0 = weak signal.}

\item{dat}{A specified dataset, if you want to simulate the loss of signal on an existing dataset, ot something.}

\item{A}{node that describes the section of the tree to be transformed (if whole tree, the root node).}
}
\description{
Simulate data over a phylogeny, or part of a phylogeny, according to the model of Munkemuller et al 2012 by simulating data with a Brownian Motion process, and then adding noise to reflect loss of signal.
}

