% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2e_sim.R
\name{simulateFlips}
\alias{simulateFlips}
\title{simuleteFlips}
\usage{
simulateFlips(attack, defense, TN = 0, suit = FALSE, afate = 0,
  dfate = 0, dtrack = NULL)
}
\arguments{
\item{attack}{Attackers attack value}

\item{defense}{Defenders resist value}

\item{TN}{Target number (for simple duel - ignores any defense stuff)}

\item{suit}{Is a specific suit needed?}

\item{dtrack}{Currently ignored}

\item{aplus}{Number of positive twists for attacker}

\item{aminus}{Number of negative twists for attacker}

\item{dplus}{Number of positive twists for defender}

\item{dminus}{Number of negative twists for defender}
}
\description{
Simulates the outcome of opposed duels.
}
