% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/smartRbind.R
\name{smartRbind}
\alias{smartRbind}
\title{smartRbind}
\usage{
smartRbind(...)
}
\arguments{
\item{A}{bunch of vectors.}
}
\description{
A function that will rbind vectors of different lengths and return a matrix, provided each vector element is named.
}
\examples{
do.call(smartRbind, list.of.vectors)
}

