% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmTree.R
\name{symmTree}
\alias{symmTree}
\title{symmTree}
\usage{
symmTree(n, bls = "equal")
}
\arguments{
\item{n}{The number of tips (must be apower of 4)}

\item{bls}{Branch lengths - either "grafen" or "equal". Defaults to "equal".}
}
\description{
Make a totally symmetrical tree of n tips with grafen branch lengths.
}
