% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDelta.R
\name{testDelta}
\alias{testDelta}
\title{testDelta}
\usage{
testDelta(tree, datafile, node = "root")
}
\arguments{
\item{tree}{The tree the data was simulated on}

\item{datafile}{The name of the datafile the simulation outputs.}

\item{node}{The node the delta is supposed to be at}
}
\description{
A function to test the output of Andrew's BM model to see if delta
has or has not been simulated using geiger to test the likelihood.
}

