% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/treeTrans.R
\name{treeTrans}
\alias{treeTrans}
\title{treeTrans}
\usage{
treeTrans(tree, param, nodes = NULL, tips = NULL, value, rescale = TRUE)
}
\arguments{
\item{tree}{A tree of class phylo}

\item{param}{The transformation applied to the tree, either "lambda", "kappa", "delta", "rate" or "EB".}

\item{nodes}{A node number or vector of nodes describing the clade(s) to be transformed.}

\item{tips}{A vector, or list of vectors, of tip labels definining clade(s) to be transformed}

\item{value}{The value or vector of values to apply to the tree or parts of the tree. Order corresponds to the order of the elements of nodes or tips.}

\item{rescale}{Whether or not to rescale the tree to have the same root-to-tip length after transformation. Defaults to TRUE.}
}
\description{
A function that transforms a tree, or a local part of a tree, according to lambda, kappa, delta, or a local rate.
}
\examples{
transTree(tree, param = "lambda", nodes = 52, value = 0)
transTree(tree, param = "rate", nodes = c(52, 91), value = 3)
transTree(tree, param = "delta", tips = list(c("dog", "cat", "moose"), c("frog", "salamander", "newt")), value = c(0.3, 2))
}
\keyword{delta}
\keyword{kappa}
\keyword{lambda}
\keyword{local}
\keyword{rates}
\keyword{transformation}
\keyword{tree}

