% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/localKappa.R
\name{localKappa}
\alias{localKappa}
\title{localKappa}
\usage{
localKappa(tree, node = NULL, tips = NULL, kappa)
}
\arguments{
\item{tree}{A tree of class phylo.}

\item{node}{A node number of vector of nodes describing the clade to be transformed.}

\item{tips}{A vector of tip names, or a list of vectors of tip names, that define the areas to be transformed. If used, node is ignored.}

\item{kappa}{The multiplier that rate is increased by in the clade specified by node.}
}
\description{
A function that transforms branch lengths according to Kappa.
}

