% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/localscalarPP.R
\name{localscalarPP}
\alias{localscalarPP}
\title{localscalarPP}
\usage{
localscalarPP(rjlog, tree, burnin = 0, thinning = 1)
}
\arguments{
\item{rjlog}{The RJ output of the run - typically suffixed with .VarRates.txt}

\item{tree}{The tree the analysis was run on}

\item{burnin}{The burnin (if required) for the mcmc (generally worked out from the other logfile)}

\item{thinning}{Thinning parameter for the MCMC output - again, worked out from the raw MCMC output logfile.}
}
\description{
A function that takes the output of a kappa, lambda, delta, VRates etc. RJ bayesTraits run and runs post-processing on it.
}

