% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/meanBranches.R
\name{meanBranches}
\alias{meanBranches}
\title{meanbranches}
\usage{
meanBranches(trees, ladderize = FALSE)
}
\arguments{
\item{trees}{An object of class multiPhylo - each tree must have the same topology.}
}
\description{
Calculate the mean branch lengths from a BayesTraits RJ analysis posterior, when topology is fixed.
}

