% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{plotShits}
\alias{plotShifts}
\alias{plotShits}
\title{plotShifts}
\usage{
plotShifts(PP, tree, scalar, scaled = FALSE, colour = "black", cex = 1,
  tips = FALSE, threshold = 0)
}
\arguments{
\item{PP}{The psotprocessor (localscalrPP) output.}

\item{tree}{The original phylogeny used for the analysis.}

\item{scalar}{The scalar to find and plot from the post processor - delta/lambda/kappa/node/branch}

\item{scaled}{Logical - plot the original tree, or the mean/sclaed tree?}

\item{colour}{The colour to use for the node circles}

\item{cex}{The scaling factor for the size of the node circles}

\item{tips}{Show tip labels?}

\item{threshold}{Threshold of probability in posterior to display deltas for, defaults to zero (i.e. shows all deltas shaded proportionally to the posterior probability)}
}
\description{
Plots the locations of the origins of scalars from the postprocessor output of bayestraits.
CURRENTLY WORKS ONLY FOR DELTAS.
}

