% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{plotShits}
\alias{plotShits}
\alias{plotShifts}
\title{plotShifts}
\usage{
plotShifts(PP, scalar, threshold = 0, colour = "red", direction = FALSE,
  scaled = "time", cex = 1, tips = FALSE, edge.cols = "black",
  edge.width = 1, main = "", scale = TRUE, bordercol = "black",
  border.width = 1, measure = "median", excludeones = FALSE,
  relativetrans = FALSE, transparency = TRUE,
  gradientcols = c("dodgerblue", "firebrick1"), rate.edges = NULL)
}
\arguments{
\item{PP}{The psotprocessor (localscalrPP) output.}

\item{scalar}{The scalar to find and plot from the post processor - delta/lambda/kappa/node/branch}

\item{threshold}{Threshold of probability in posterior to display deltas for, defaults to zero (i.e. shows all deltas shaded proportionally to the posterior probability)}

\item{colour}{The colour to use for the node circles}

\item{direction}{If TRUE nodes are identified with an upward triangle if the parameter is > 1 and a downward if < 1}

\item{scaled}{Plot the original tree (scaled = "time", the default), or the mean/sclaed tree (scaled = "mean") or plot the tree scaled only by scalars present above the threshold (scaled = "threshold")?}

\item{cex}{The scaling factor for the size of the node circles}

\item{tips}{Show tip labels?}

\item{edge.cols}{A vector of edge colours for the phylogeny. Ignored if plotting rates, or using rates.edges}

\item{edge.width}{The thicknes of the edges of the plotted tree.}

\item{main}{Title for the plot}

\item{scale}{Include scale bar?}

\item{bordercol}{What colour do you want the border of the node shapes to be? Defaults to black.}

\item{border.width}{Width of the border of the node shapes.}

\item{measure}{When plotting "siginficant" tree, what measure of the parameter? Median (default), mode or mean.}

\item{relativetrans}{If TRUE (defaults to FALSE) the scale of transparency will go from the threshold (totally transparent) to the maximum presence (full opacity).}

\item{transparency}{Plot the node labels according to their presnce in the posterior?}

\item{gradientcols}{A vector of two colours - the min and max colours used when colouring the tree according to percentage time rate scaled (when threshold = 0) or using rate.edges.}

\item{rate.edges}{Takes a numeric value between 0 and 1. If NULL (default) then node shapes are plotted for a transformation. If equal to zero then node shapes are plotted along with a colour gradient on the branches for rates (if in the posterior), and if set to a threshold then the branches are coloured black/red for whether there is a scalar over the threshold (red) along with node scalars.}

\item{exludeones}{If plotting according to a threshold of significance, should 1s (i.e. no scalar) be excluded from the posterior when calculating average scalar?}
}
\description{
Plots the locations of the origins of scalars from the postprocessor output of bayestraits.
CURRENTLY WORKS ONLY FOR DELTAS.
}
