% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simuDatEb.R
\name{simuDatEB}
\alias{simuDatEB}
\title{simuDatEb}
\usage{
simuDatEB(tree, node, sig, a, t, decay = "exp", dat = NULL)
}
\arguments{
\item{tree}{}

\item{node}{}

\item{sig}{The standard deviation of the random component of the BM simulation. NOT sigma squared.}

\item{a}{Initial rate increase before decay. If using the power decay this is, essentially, delta.}

\item{t}{The number of half lives for the decay to take place over. Essentially a high number means the rate decays rapidly, and a low one means it decays slowly. This is used rather than a rate parameter, since decay is relative to branch lengths, but half lives are not.}

\item{decay}{Either "exp" (exponential) or "pwr" (power). The shape of the decay from the initial rate.}

\item{dat}{If you want to use some data you already have and change it this way, put it here.}
}
\description{
Simulate data over a phylogeny, or sub section of a phylogeny, in order to reflect variable rates.
Based on adding a constant to the tip data branch by branch and weighting by branch length and does
not need to stretch the phylogeny.
}

