% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simuDatRate.R
\name{simuDatRate}
\alias{simuDatRate}
\title{simuDatRate}
\usage{
simuDatRate(tree, node, a, sig, dat = NULL)
}
\arguments{
\item{tree}{}

\item{node}{}

\item{a}{The constant that is weighted by branch length and added to the data at the tips. Once the relationship between scalar and this is known, this can be scalar}

\item{sig}{The standard deviation of the random component of the BM simulation. NOT sigma squared.}

\item{dat}{If you want to use some data you already have and change it this way, put it here.}
}
\description{
Simulate data over a phylogeny, or sub section of a phylogeny, in order to reflect variable rates.
Based on adding a constant to the tip data branch by branch and weighting by branch length and does
not need to stretch the phylogeny.
}

